Rename ViewGroupProfile -> ViewGroupPage
