+++ metaspace/webapp/src/modules/GroupProfile/ViewGroupPage.spec.ts
@@ -1,11 +1,11 @@
 import { mount, Stubs } from '@vue/test-utils';
 import Vue from 'vue';
-import ViewGroupProfile from './ViewGroupProfile.vue';
+import ViewGroupPage from './ViewGroupPage.vue';
 import router from '../../router';
 import { initMockGraphqlClient, provide } from '../../../tests/utils/mockGraphqlClient';
 
 
-describe('ViewGroupProfile', () => {
+describe('ViewGroupPage', () => {
   const mockMembersForAdmins = [
     {
       role: 'GROUP_ADMIN',
@@ -78,7 +78,7 @@ describe('ViewGroupProfile', () => {
     it('should match snapshot (non-member)', async () => {
       initMockGraphqlClient(graphqlMocks);
       const maxVisibleDatasets = 3;
-      const wrapper = mount(ViewGroupProfile, { router, stubs, provide, sync: false });
+      const wrapper = mount(ViewGroupPage, { router, stubs, provide, sync: false });
       wrapper.setData({ maxVisibleDatasets }); // Also test that the datasets list is correctly clipped
       await Vue.nextTick();
 
@@ -99,7 +99,7 @@ describe('ViewGroupProfile', () => {
 
     it('should match snapshot (non-member)', async () => {
       initMockGraphqlClient(graphqlMocks);
-      const wrapper = mount(ViewGroupProfile, { router, stubs: stubsWithMembersList, provide, sync: false });
+      const wrapper = mount(ViewGroupPage, { router, stubs: stubsWithMembersList, provide, sync: false });
       await Vue.nextTick();
 
       expect(wrapper).toMatchSnapshot();
@@ -108,7 +108,7 @@ describe('ViewGroupProfile', () => {
     it('should match snapshot (invited)', async () => {
       mockGroupFn.mockImplementation(() => ({...mockGroup, currentUserRole: 'INVITED'}));
       initMockGraphqlClient(graphqlMocks);
-      const wrapper = mount(ViewGroupProfile, { router, stubs: stubsWithMembersList, provide, sync: false });
+      const wrapper = mount(ViewGroupPage, { router, stubs: stubsWithMembersList, provide, sync: false });
       await Vue.nextTick();
 
       expect(wrapper).toMatchSnapshot();
@@ -117,7 +117,7 @@ describe('ViewGroupProfile', () => {
     it('should match snapshot (member)', async () => {
       mockGroupFn.mockImplementation(() => ({...mockGroup, currentUserRole: 'MEMBER', members: mockMembersForMembers}));
       initMockGraphqlClient(graphqlMocks);
-      const wrapper = mount(ViewGroupProfile, { router, stubs: stubsWithMembersList, provide, sync: false });
+      const wrapper = mount(ViewGroupPage, { router, stubs: stubsWithMembersList, provide, sync: false });
       await Vue.nextTick();
 
       expect(wrapper).toMatchSnapshot();
@@ -126,7 +126,7 @@ describe('ViewGroupProfile', () => {
     it('should match snapshot (manager, including table)', async () => {
       mockGroupFn.mockImplementation(() => ({...mockGroup, currentUserRole: 'GROUP_ADMIN', members: mockMembersForAdmins}));
       initMockGraphqlClient(graphqlMocks);
-      const wrapper = mount(ViewGroupProfile, { router, stubs, provide, sync: false });
+      const wrapper = mount(ViewGroupPage, { router, stubs, provide, sync: false });
       await Vue.nextTick();
 
       expect(wrapper).toMatchSnapshot();
@@ -141,7 +141,7 @@ describe('ViewGroupProfile', () => {
     it('should match snapshot', async () => {
       mockGroupFn.mockImplementation(() => ({...mockGroup, currentUserRole: 'GROUP_ADMIN'}));
       initMockGraphqlClient(graphqlMocks);
-      const wrapper = mount(ViewGroupProfile, { router, stubs, provide, sync: false });
+      const wrapper = mount(ViewGroupPage, { router, stubs, provide, sync: false });
       await Vue.nextTick();
 
       expect(wrapper).toMatchSnapshot();
@@ -153,7 +153,7 @@ describe('ViewGroupProfile', () => {
     mockGroupFn.mockImplementation(() => ({...mockGroup, urlSlug}));
     initMockGraphqlClient(graphqlMocks);
 
-    const wrapper = mount(ViewGroupProfile, { router, stubs, provide, sync: false });
+    const wrapper = mount(ViewGroupPage, { router, stubs, provide, sync: false });
     await Vue.nextTick();
 
     expect(router.currentRoute.params.groupIdOrSlug).toEqual(urlSlug);
