+++ metaspace/webapp/src/modules/GroupProfile/index.ts
@@ -1,4 +1,4 @@
-export {default as ViewGroupProfile} from './ViewGroupProfile.vue'
+export {default as ViewGroupPage} from './ViewGroupPage.vue'
 export {default as EditGroupProfile} from './GroupSettings.vue'
 export {default as CreateGroupPage} from './CreateGroupPage.vue'
 export {default as TransferDatasetsDialog} from './TransferDatasetsDialog.vue'
